requirements.txt
PySide6==6.10.1
numpy==2.2.2
Pillow==11.1.0
reportlab==4.3.0
colour-science==0.4.6
